pyproject.toml
[project]
name = "learning-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic-settings",
  "python-dotenv",
  "pyyaml",
  "httpx",
  "pypdf",
  "qdrant-client",
  "fastapi-mcp",
  "mcp[client]>=1.14.0",
  "autogen-agentchat>=0.2.0",
  "autogen-ext>=0.2.0",
  "openai>=1.40.0",
    "tiktoken",
  
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
